pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-audit"
version = "0.1.0"
description = "Django-based toolkit for auditing ML predictions in production."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Rajat Jog", email = "rajatjog1294@gmail.com" },
]
dependencies = [
  "django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ml-audit"

[tool.setuptools.packages.find]
where = ["src"]
